fix: use OUTPUT INTO on SqlServer for returning columns (#5361)
* test: Add test for GitHub issue #5160

* fix: use OUTPUT INTO on SqlServer for returning columns

Closes #5160

* chore: fix typo in comment

* fix: use full types with length e.g. nvarchar(255)

* refactor: move output table declaration to driver

* fix: don't try to drop temporary tables

+++ src/driver/sqlserver/SqlServerQueryRunner.ts
@@ -1370,6 +1370,11 @@ export class SqlServerQueryRunner extends BaseQueryRunner implements QueryRunner
                 return Promise.all(dropFkQueries.map(result => result["query"]).map(dropQuery => this.query(dropQuery)));
             }));
             await Promise.all(allTablesResults.map(tablesResult => {
+                if (tablesResult["TABLE_NAME"].startsWith("#")) {
+                    // don't try to drop temporary tables
+                    return;
+                }
+
                 const dropTableSql = `DROP TABLE "${tablesResult["TABLE_CATALOG"]}"."${tablesResult["TABLE_SCHEMA"]}"."${tablesResult["TABLE_NAME"]}"`;
                 return this.query(dropTableSql);
             }));
